Change NijMatrix to long long unsigned int as well.

+++ scripts/critchlow-metrics.h
@@ -10,7 +10,7 @@
 typedef struct nij_matrix_s {
 	size_t numRanks;
 	
-	size_t *nijs;
+	long long unsigned int *nijs;
 } NijMatrix;
 
 extern void freeNijMatrix(NijMatrix *nij);
